+++ src/services/helpers.ts
@@ -45,14 +45,21 @@ export async function findAllGeneric<T>(
   options: OptionsWithRql | undefined,
   level = 1
 ): Promise<T[]> {
-  if (level === 1 && options?.rql && options.rql.includes('limit(')) {
-    throw new Error('Do not pass in limit operator with findAll');
+  if (level === 1 && options?.rql) {
+    if (options.rql.includes('limit(')) {
+      throw new Error('Do not pass in limit operator with findAll');
+    }
+
+    if (options.rql.includes('skipCount(')) {
+      throw new Error('Do not pass in skipCount operator with findAll');
+    }
   }
 
   // return async options => {
   // Extra check is needed because this function is call recursively with updated RQL
   // But on the first run, we need to set the limit to the max to optimize
   const result: PagedResult<T> = await find({
+    ...options,
     rql:
       options?.rql && options.rql.includes('limit(') ?
         options.rql :
@@ -71,6 +78,7 @@ export async function findAllGeneric<T>(
       ...(await findAllGeneric(
         find,
         {
+          ...options,
           rql: rqlBuilder(options?.rql)
             .limit(result.page.limit, result.page.offset + result.page.limit)
             .build(),
@@ -90,6 +98,7 @@ export function addPagersFn<T>(
 
   async function previous() {
     result = await find({
+      ...options,
       rql: rqlBuilder(options?.rql)
         .limit(
           result.page.limit,
@@ -102,6 +111,7 @@ export function addPagersFn<T>(
 
   async function next() {
     result = await find({
+      ...options,
       rql: rqlBuilder(options?.rql)
         .limit(
           result.page.limit,